Basic auto-completion of object members
Add support for rudimentary auto-completion of object members.

Nim allows one to call any procedure with a compatible first argument
using the method call syntax, this means that in order to auto-complete
proc names we have to search _all_ project files (and in the future also
all libraries) for matching procs, which is not feasible with the
current word-based index (it might be doable with stub-based indices).
Alternatively, we could resolve all the imported modules and only search
those, but then we wouldn't be able to automatically add imports when a
symbol from a non-imported module is used (a feature that we'd like to
have some day).

I ended up with a reserved approach, whereby auto-completion only
includes the names of those procs that are declared in the current file
(the one where auto-completion is invoked) as well as those declared in
the file that contains the object type definition.